Add use of tag dictionary. Add final output stage. Switch to doubles instead of floats. Other changes.

+++ Assignment3/src/MapReduce/ContextInfoReducer.java
@@ -13,7 +13,7 @@ import java.util.ArrayList;
 // import org.apache.commons.logging.Log;
 // import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.mapreduce.Reducer;
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.LongWritable;
 
@@ -21,7 +21,7 @@ import org.apache.hadoop.io.LongWritable;
  *
  * @author yarneo
  */
-public class ContextInfoReducer extends Reducer<Text,Pair,Text,FloatWritable> {
+public class ContextInfoReducer extends Reducer<Text,Pair,Text,DoubleWritable> {
     // The Karmasphere Studio Workflow Log displays logging from Apache Commons Logging, for example:
     // private static final Log LOG = LogFactory.getLog("MapReduce.ContextInfoReducer");
 
@@ -36,9 +36,9 @@ public class ContextInfoReducer extends Reducer<Text,Pair,Text,FloatWritable> {
         for(Pair val : valueList) {
         	//Key: word /tab context   Value: count(word in context) / count(context) = p(word|context)
         	String outStr = val.getText().toString() + "-,-" + key.toString();
-        	float outFloat = ((float)val.getNum().get() / count_of_context);
+        	double outFloat = ((double)val.getNum().get() / count_of_context);
 //        	System.out.println("CONTEXTINFOREDUCE: " + outStr + " " + outFloat);
-        	context.write(new Text(outStr), new FloatWritable(outFloat));
+        	context.write(new Text(outStr), new DoubleWritable(outFloat));
         }
     }
 }
